+++ backend/package.json
@@ -10,8 +10,11 @@
   },
   "dependencies": {
     "@prisma/client": "^6.6.0",
+    "axios": "^1.9.0",
     "bcrypt": "^5.1.1",
+    "dotenv": "^16.5.0",
     "express": "^5.1.0",
+    "mysql2": "^3.14.1",
     "next": "15.2.4",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
@@ -21,7 +24,8 @@
     "@tailwindcss/postcss": "^4",
     "@types/bcrypt": "^5.0.2",
     "@types/express": "^5.0.1",
-    "@types/node": "^20",
+    "@types/mysql": "^2.15.27",
+    "@types/node": "^20.17.32",
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "eslint": "^9",
