+++ backend/tsconfig.json
@@ -8,8 +8,10 @@
     "noEmit": true,
     "esModuleInterop": true,
     "module": "esnext",
-    "moduleResolution": "bundler",
+    "moduleResolution": "node",
     "resolveJsonModule": true,
+    "typeRoots": ["node_modules/@types"],
+    "types": ["node"],
     "isolatedModules": true,
     "jsx": "preserve",
     "incremental": true,
